requirements.txt
pandas
numpy
joblib
requests
azureml-core
azure-cli
azureml-sdk[cli]
pytest
pytest-cov
sklearn